Refactoring entier de l'app
