+++ Project02Eclipse/src/com/hemebiotech/analytics/Main.java
@@ -1,30 +1,28 @@
 package com.hemebiotech.analytics;
 
+import com.hemebiotech.analytics.counter.AnalyticsCounter;
+import com.hemebiotech.analytics.reader.ReadSymptomDataFromFile;
+import com.hemebiotech.analytics.writer.WriteSymptomDataOnFile;
+
 import java.util.List;
 import java.util.Map;
 
-/** 
+/**
  * Run the Application
- * 
  */
 public class Main {
-    
+
     public static void main(String[] args) {
-        
-        ReadSymptomDataFromFile readSymptomDataFromFile = new ReadSymptomDataFromFile("/Users/ray/Documents/DeveloppeurApplicationJAVA/P2_debuggez_une_application_Java/Project_DA_Java_EN_Come_to_the_Rescue_of_a_Java_Application/Project02Eclipse/symptoms.txt");
+        // Reading symptoms from file
+        ReadSymptomDataFromFile readSymptomDataFromFile = new ReadSymptomDataFromFile("symptoms.txt");
+        List<String> listOfSymptoms = readSymptomDataFromFile.getSymptoms();
 
-		List<String> listOfSymptoms = readSymptomDataFromFile.getSymptoms();
-
-        // listOfSymptom.forEach((symptom) -> System.out.println(symptom));
-
+        // Count symptoms from List
         AnalyticsCounter analyticsCounter = new AnalyticsCounter(listOfSymptoms);
-
         Map<String, Integer> symptomCounter = analyticsCounter.symptomCounter();
 
-        // symptomCounter.forEach((symptom, count) -> System.out.println(symptom + "=" + count));
-
-        WriteSymptomDataOnFile symptomsCountFile = new WriteSymptomDataOnFile(symptomCounter, "/Users/ray/Documents/DeveloppeurApplicationJAVA/P2_debuggez_une_application_Java/Project_DA_Java_EN_Come_to_the_Rescue_of_a_Java_Application/Project02Eclipse/result.out");
-
+        // write symptom's count on file
+        WriteSymptomDataOnFile symptomsCountFile = new WriteSymptomDataOnFile(symptomCounter, "result.out");
         symptomsCountFile.writeSymptoms();
 
     }
