+++ Project02Eclipse/src/com/hemebiotech/analytics/writer/ISymptomWriter.java
@@ -1,16 +1,15 @@
-package com.hemebiotech.analytics;
+package com.hemebiotech.analytics.writer;
 
-/** 
+/**
  * Anything that will write symptom data on a source
  * The important part is, the operation get symptom data from a Map,
  * sort in alphabetic order
- * 
  */
 public interface ISymptomWriter {
-    /** 
+    /**
      * Write an oredered Map of every symptom with his count, on the source's file,
-     * without any duplicate 
+     * without any duplicate
      */
-    void writeSymptoms ();
+    void writeSymptoms();
 
 }
