+++ Project02Eclipse/src/com/hemebiotech/analytics/writer/WriteSymptomDataOnFile.java
@@ -1,4 +1,4 @@
-package com.hemebiotech.analytics;
+package com.hemebiotech.analytics.writer;
 
 import java.io.BufferedWriter;
 import java.io.FileWriter;
@@ -9,20 +9,18 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-/** 
+/**
  * Sort and write symptoms data on a source
- * 
  */
-public class WriteSymptomDataOnFile implements ISymptomWriter{
+public class WriteSymptomDataOnFile implements ISymptomWriter {
 
     private Map<String, Integer> symptomMap;
     private String filepath;
 
     /**
-	 * 
      * @param symptomMap a map of every symptom with their count from the source with no duplicate
-	 * @param filepath a full or partial path to file with symptom strings in it, one per line
-	 */
+     * @param filepath   a full or partial path to file with symptom strings in it, one per line
+     */
     public WriteSymptomDataOnFile(Map<String, Integer> symptomMap, String filepath) {
         this.symptomMap = symptomMap;
         this.filepath = filepath;
@@ -32,14 +30,14 @@ public class WriteSymptomDataOnFile implements ISymptomWriter{
     public void writeSymptoms() {
 
         Map<String, Integer> orderedSymptomsCount = orderSymptomsCount();
-        
-        if(filepath != null) {
+
+        if (filepath != null) {
             try {
-                BufferedWriter writer = new BufferedWriter (new FileWriter(filepath, false));
+                BufferedWriter writer = new BufferedWriter(new FileWriter(filepath, false));
 
                 orderedSymptomsCount.forEach((symptom, count) -> {
                     try {
-                        writer.write(symptom +"="+count);
+                        writer.write(symptom + "=" + count);
                         writer.newLine();
                     } catch (IOException e) {
                         e.printStackTrace();
@@ -53,7 +51,6 @@ public class WriteSymptomDataOnFile implements ISymptomWriter{
     }
 
     /**
-     * 
      * @return an ordered alphbetic map of symtoms with their count, no duplicates are possible/probable
      */
     private Map<String, Integer> orderSymptomsCount() {
